ground-station/requirements.txt
aprslib==0.7.2
PyYAML==6.0
